pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "ErmisCho",email = "ermis.chorinopoulos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "langchain (>=1.1.3,<2.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "tavily-python (>=0.7.15,<0.8.0)",
    "chromadb (>=1.3.7,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "langchain-core (>=1.2.1,<2.0.0)",
    "langchain-tavily (>=0.2.14,<0.3.0)",
    "langchain-google-genai (>=4.0.0,<5.0.0)",
    "langchain-pinecone (>=0.2.13,<0.3.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "-ra"
filterwarnings = [
  'ignore:Field name "output_schema" in "TavilyResearch" shadows an attribute in parent "BaseTool":UserWarning:langchain_tavily.*',
  'ignore:Field name "stream" in "TavilyResearch" shadows an attribute in parent "BaseTool":UserWarning:langchain_tavily.*',
]